% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binarizeAndCombine.R
\name{binarizeAndCombine}
\alias{binarizeAndCombine}
\title{Binarize images and combine to a single binary image}
\usage{
binarizeAndCombine(
  im,
  method = c("tpoint", "geometric", "peak"),
  binMatrix = NULL,
  ...
)
}
\arguments{
\item{im}{matrix or array representing a image or a stack of images}

\item{method}{character, can be "tpoint", "geometric", "peak"}

\item{binMatrix}{matrix, a binary matrix. Overwrites method and uses this matrix to binarize the data.}
}
\value{
matrix with binary image
}
\description{
Binarize images and combine to a single binary image
}
