% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineIonImages.R
\name{combineIonImages}
\alias{combineIonImages}
\title{Combine two stacks ion images into one stack of ion images}
\usage{
combineIonImages(ii1, ii2, add = TRUE)
}
\arguments{
\item{ii1}{Array, ion images as returend by MSISclices from MALDIquant}

\item{ii2}{Array, ion images as returend by MSISclices from MALDIquant}

\item{add}{Logical, if the dimensions of ii1 and ii2 dont match, add rows and cols to compensate}
}
\value{
A stack of combined ion images
}
\description{
This function can be used to combine different stacks of ion images (e.g. from different polarities).
For this to work properly the images have to be aligned.
Small differences (e.g. one additional row) can be compensated with this function.
Be sure you know what you are doing and check if the quality of the result is sufficent for your means!
}
