% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometricThreshold.R
\name{geometricThreshold}
\alias{geometricThreshold}
\title{Unimodal threshold based on geometric method}
\usage{
geometricThreshold(intensities, breaks = 500, plot = FALSE)
}
\arguments{
\item{intensities}{Vector of intensities.}

\item{breaks}{Number of bins of the histogram.}

\item{plot}{Plots the internally generated plot.}
}
\value{
Threshold value (see description)
}
\description{
Calculates a threshold for unimodal distribuated values based on histogram thresholding.
Implementation of Rosin, P. L. (2001). "Unimodal thresholding." Pattern Recognition 34(11): 2083-2096.
A line is drawn between the bin with the highest frequency and the bin with the highest intensity/value (main line).
After for each bin another line, orthogonal to the first line, is drawn to the "top" of each bin.
The value of the bin with the longst distance to the main line is defined as threshold
}
