% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_IdxFromID.R
\name{get_IdxFromID}
\alias{get_IdxFromID}
\title{Get indicies of spectra per plaque ID}
\usage{
get_IdxFromID(pp, ID, mzValue = "unified")
}
\arguments{
\item{pp}{plaque picker result list. See \code{plaquePicker}}

\item{ID}{vector of ID, set to NA to return all Indicies associated with any clump.}

\item{mzValue}{mzValue to ID is refering to. By default the unified IDs are used but this can also be overwritten by entering a specific mz.}
}
\value{
vector of indicies refering to the spectra in the MALDIquant spectra list used to construct the ion images used.
}
\description{
Get indicies of spectra per plaque ID
}
