% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plaquePicker.R
\name{get_intensities}
\alias{get_intensities}
\title{Get intensity values from ion images}
\usage{
get_intensities(comp, ii)
}
\arguments{
\item{comp}{matrix, locations of connected components}

\item{ii}{array, ion images}
}
\value{
list of intensities.
}
\description{
Get intensity values from ion images
}
