% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mzValues.R
\name{get_mzValues}
\alias{get_mzValues}
\title{get mz values in plaquePicker result list}
\usage{
get_mzValues(pp)
}
\arguments{
\item{pp}{plaquePicker result list}
}
\value{
vector of mz values as character.
}
\description{
get mz values in plaquePicker result list
}
