% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plaquePicker.R
\name{get_specIdx}
\alias{get_specIdx}
\title{Internal helper function to extract spectra indices}
\usage{
get_specIdx(comp, coord)
}
\arguments{
\item{comp}{matrix, result of \code{raster::clump}. Matrix with indicies for each unique connected component.}

\item{coord}{array of coordinates as returned by \code{MALDIquant::coordinates()}.}
}
\value{

}
\description{
Internal helper function to extract spectra indices
}
