% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NLGF67w_mouse1_rep1}
\alias{NLGF67w_mouse1_rep1}
\title{Ion images from a NLGF mouse at age 67 weeks}
\format{
Array of ion images from NLGF mouse at 67 of weeks from Enzlein, et. al. 2020. mz = 4059.9, 4159.1, 4442.6 +-5 Da.
}
\usage{
NLGF67w_mouse1_rep1
}
\description{
Ion images from a NLGF mouse at age 67 weeks
}
\keyword{datasets}
