% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NLGF67w_mouse1_rep1_coord}
\alias{NLGF67w_mouse1_rep1_coord}
\title{Coordinates from a NLGF mouse at age 67 weeks}
\format{
Double, coordinates as extracted by \code{MALDIquant::coordinates} from NLGF mouse at 67 of weeks from Enzlein, et. al. 2020
}
\usage{
NLGF67w_mouse1_rep1_coord
}
\description{
Coordinates from a NLGF mouse at age 67 weeks
}
\keyword{datasets}
