% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NLGF67w_mouse1_rep1_spec}
\alias{NLGF67w_mouse1_rep1_spec}
\title{Processed and reduced spectra from a NLGF mouse at age 67 weeks}
\format{
list of \code{MALDIquant::MassSpectrum} from NLGF mouse at 67 of weeks from Enzlein, et. al. 2020. Spectra were reduced (mz 4k-4.5k, 25% data points) and processed (TIC norm, SavitzkyGolay smoothed, baseline removed (TopHat))
}
\usage{
NLGF67w_mouse1_rep1_spec
}
\description{
Processed and reduced spectra from a NLGF mouse at age 67 weeks
}
\keyword{datasets}
