% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_venn.R
\name{plot_venn}
\alias{plot_venn}
\title{Generate a Venn-diagram for a plaquePicker result list}
\usage{
plot_venn(
  pp,
  mzIdx = 1:3,
  mzNames = NULL,
  plot = TRUE,
  method = c("plaque", "pixel"),
  relative = TRUE
)
}
\arguments{
\item{pp}{plaquePicker result list}

\item{mzIdx}{integer vector, indicies of mz values. To get an overview of mz values try \code{get_mzValues()}. Can not be longer then 3.}

\item{mzNames}{character vector, names of mz values (e.g. "ABeta1-38"). If set to NULL (default) the mz values themself will be used as names during plotting.}

\item{plot}{logical, if set to FALSE the \code{Venn} object is returned.}

\item{method}{character, if set to "plaque" the Venn diagram will be based on overlapping signals per plaque. If set to "pixel" then it will be pixel-wise.}

\item{relative}{logical, if set to false the number of plaques is returned instead of the percentage.}
}
\description{
Generate a Venn-diagram for a plaquePicker result list
}
