% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometricThreshold.R
\name{point2LineDist}
\alias{point2LineDist}
\title{point2LineDist}
\usage{
point2LineDist(p1, p2, x)
}
\arguments{
\item{p1}{Point 1 c(x,y)}

\item{p2}{Point 2 c(x,y)}

\item{x}{Point which distance should be calculated in relation to line formed by \code{p1} and \code{p2}.}
}
\value{
Distance of \code{x} to line defined by \code{p1} and \code{p2}
}
\description{
Calculates the distance between point \code{x} to the line defined by \code{p1} and \code{p2}.
ref: https://en.wikipedia.org/wiki/Distance_from_a_point_to_a_line
}
