% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcor.R
\name{qcor}
\alias{qcor}
\title{Quantile correction}
\usage{
qcor(im, q = 0.999)
}
\arguments{
\item{im}{matrix containing intensities of ion image.}

\item{q}{quantile that is set to 100% intensity}
}
\value{
matrix of quantile corrected intensities.
}
\description{
Function to perform quantile correction for better visual representation of ion images.
}
\examples{
im_cor <- qcor(NLGF67w_mouse1_rep1[,,1], 0.9995)
}
