% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpoint.R
\name{tpoint}
\alias{tpoint}
\title{T-Point thresholding}
\usage{
tpoint(ints, breaks = 500, diagnosis = FALSE, plot = FALSE)
}
\arguments{
\item{ints}{numeric vector of intensities.}

\item{breaks}{numeric, number of bins in histogram.}

\item{diagnosis}{logical, return set of internal values additonal to threshold}

\item{plot}{logical, show plot for diagnosis.}
}
\value{
numeric threshold.
}
\description{
implemation described in Coudray et. al. 2010,
see https://www.sciencedirect.com/science/article/pii/S0167865509003675?via%3Dihub#fd6
}
\examples{
thresh <- tpoint(as.vector(NLGF67w_mouse1_rep1[,,1]))
}
